clusterblast: move diamond db checks to common.subprocessing.diamond
