+++ antismash/common/subprocessing/test/data/format0.dmnd
