+++ antismash/common/subprocessing/test/data/format1.dmnd
