+++ antismash/common/subprocessing/test/integration_subprocessing.py
@@ -8,8 +8,11 @@ import os
 import time
 import unittest
 
-from antismash.config import build_config, destroy_config, get_config
-from antismash.common import subprocessing
+from helperlibs.wrappers.io import TemporaryDirectory
+
+from antismash.common import subprocessing, path
+from antismash.config import build_config, destroy_config, get_config, update_config
+from antismash.main import get_all_modules
 
 
 def dummy(value=None):
@@ -153,3 +156,33 @@ class TestParallelExecute(unittest.TestCase):
             subprocessing.parallel_execute([["sleep", "10"]]*2, timeout=1)
         elapsed = time.time() - start
         assert elapsed < 1.5
+
+
+class TestDiamondDatabaseChecks(unittest.TestCase):
+    def setUp(self):
+        self.format0_file = path.get_full_path(__file__, "data", "format0.dmnd")
+        self.format1_file = path.get_full_path(__file__, "data", "format1.dmnd")
+        self.empty = path.get_full_path(__file__, "data", "empty.dmnd")
+
+        options = build_config([], isolated=True, modules=get_all_modules())
+        self.old_config = get_config().__dict__
+        self.options = update_config(options)
+
+    def tearDown(self):
+        destroy_config()
+        update_config(self.old_config)
+
+    def test_check_diamond_db_compatible(self):
+        with TemporaryDirectory(change=True):
+            dummy_fasta = "dummy.fa"
+            dummy_db = "dummy.dmnd"
+            with open(dummy_fasta, "w") as handle:
+                handle.write(">test\nM\n")
+            subprocessing.run_diamond_makedb(dummy_db, dummy_fasta)
+            compatible_format = subprocessing.diamond._extract_db_format(dummy_db)
+            assert subprocessing.diamond.check_diamond_db_compatible(dummy_db)
+
+        broken_file = self.format0_file if compatible_format > 0 else self.format1_file
+
+        assert not subprocessing.diamond.check_diamond_db_compatible(broken_file)
+        assert not subprocessing.diamond.check_diamond_db_compatible(self.empty)
